+++ src/app/api/invoices/[id]/credit-note/route.ts
@@ -1,11 +1,11 @@
-import { NextRequest, NextResponse } from 'next/server'
-import { db } from '@/lib/db'
-import { z } from 'zod'
-import { logger } from '@/lib/logger'
-import { formatInvoiceNumber } from '@/lib/invoice-utils'
-import { decryptField } from '@/lib/field-encryption'
 import { requireStoreAccess } from '@/lib/api-auth'
-import { getSoftwareProviderNIT, DIAN_CONSUMIDOR_FINAL_NIT } from '@/lib/constants'
+import { DIAN_CONSUMIDOR_FINAL_NIT, getSoftwareProviderNIT } from '@/lib/constants'
+import { db } from '@/lib/db'
+import { decryptField } from '@/lib/field-encryption'
+import { formatInvoiceNumber } from '@/lib/invoice-utils'
+import { logger } from '@/lib/logger'
+import { NextRequest, NextResponse } from 'next/server'
+import { z } from 'zod'
 
 export const dynamic = 'force-dynamic'
 
@@ -26,7 +26,7 @@ const creditNoteFromInvoiceSchema = z.object({
   // Items afectados (opcional — si no se proporciona, es una nota global)
   items: z.array(z.object({
     description: z.string().max(500),
-    quantity: z.number().int().positive(),
+    quantity: z.number().positive(),
     unitPrice: z.number().int().min(0),
     taxCode: z.string().max(10).optional().default('01'),
     taxRate: z.number().int().min(0).max(100).optional().default(0),
